+++ src/contexts/index.js
@@ -1,0 +1,2 @@
+export { useAuth, AuthProvider } from "./auth-context";
+export { useData, DataProvider } from "./data-context";
