+++ src/index.js
@@ -4,8 +4,7 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter } from "react-router-dom";
-import { DataProvider } from "./contexts/data-context";
-import { AuthProvider } from "./contexts/auth-context";
+import { DataProvider, AuthProvider } from "./contexts";
 // Call make Server
 makeServer();
